+++ src/main/java/com/kh/semi/order/model/dao/OrderDao.java
@@ -1,5 +1,13 @@
 package com.kh.semi.order.model.dao;
 
-public class OrderDao {
+import java.util.List;
+
+import com.kh.semi.order.model.vo.Order;
+
+public interface OrderDao {
+
+	int insertOrder(Order o);
+
+	List<Order> selectOrderList();
 
 }
