+++ src/main/java/com/kh/semi/order/model/vo/Order.java
@@ -1,4 +1,4 @@
-package com.kh.semi.Orders.model.vo;
+package com.kh.semi.order.model.vo;
 
 import java.sql.Date;
