--- src/main/java/com/kh/semi/order/model/vo/Orders.java
@@ -1,5 +1,0 @@
-package com.kh.semi.order.model.vo;
-
-public class Orders {
-
-}
